Changes to main

+++ app/src/main/java/com/example/dictionaryapp/screens/home/ui/MainActivity.kt
@@ -54,7 +54,7 @@ class MainActivity : AppCompatActivity() {
         val result = convertStringToArray(jsonData)
 
         result.mapIndexed { index, s ->
-            viewModelHome.addList(Word(s, index))
+            viewModelHome.addList(Word(s, index.toLong()))
         }
 
     }
